issue #364 - fix smaller issues found in tests
fix sim connection
fix java nCopies
fix java random of list
fix cpp random of list
fix rfid wait until
fix nxt display bool
fix botnroll ternary
fix botnroll show text colour
fix edison function return
fix unowifirev2 pin write

+++ RobotArdu/src/test/resources/ast/actions/arduino_analog_digital_output_test.ino
@@ -15,6 +15,6 @@ void setup()
 
 void loop()
 {
-    digitalWrite(_output_A, 1);
-    analogWrite(_output_A2, 1);
+    digitalWrite(_output_A, (int) 1);
+    analogWrite(_output_A2, (int) 1);
 }
